+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* Added support for asynchronous back-ends (experimental).
 * Support for back-end command `crop` dropped (it was never documented).
 * Support for Emacs 23 dropped.
 * New user option `company-abort-manual-when-too-short`.